refactor: per-module unit tests in prosemirror search
